requirements.txt
# Internet Adoption Analysis — Python dependencies
# Minimum Python version: 3.10
#
# Install with:  pip install -r requirements.txt
# All versions are pinned to stable releases known to be compatible as of early 2025.

# ---------------------------------------------------------------------------
# HTTP & retry
# ---------------------------------------------------------------------------
requests==2.32.3
tenacity==9.0.0

# ---------------------------------------------------------------------------
# Data manipulation & numerical computing
# ---------------------------------------------------------------------------
pandas==2.2.3
numpy==1.26.4

# ---------------------------------------------------------------------------
# Machine learning (k-means, PCA, silhouette score, StandardScaler)
# ---------------------------------------------------------------------------
scikit-learn==1.5.2

# ---------------------------------------------------------------------------
# Visualisation
# ---------------------------------------------------------------------------
matplotlib==3.9.4
seaborn==0.13.2
Pillow==11.1.0

# ---------------------------------------------------------------------------
# Statistics
# ---------------------------------------------------------------------------
scipy==1.13.1

# ---------------------------------------------------------------------------
# Configuration parsing (YAML)
# ---------------------------------------------------------------------------
pyyaml==6.0.2

# ---------------------------------------------------------------------------
# Readability check (Flesch-Kincaid)
# ---------------------------------------------------------------------------
textstat==0.7.4

# ---------------------------------------------------------------------------
# Country code utilities (ISO3 → ISO2 mapping)
# ---------------------------------------------------------------------------
pycountry==24.6.1

# ---------------------------------------------------------------------------
# Testing
# ---------------------------------------------------------------------------
pytest==8.3.5
hypothesis==6.130.5
